Add test coverage for rendering widget
